requirements.txt
# Core Dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.0.0

# AI & API
google-generativeai>=0.8.0
python-dotenv>=1.0.0

# Text Processing
emoji==1.4.2

# Optional: For development only (not needed for deployment)
# transformers>=4.30.0
# torch>=2.0.0
# arabert>=1.0.0
# camel-tools>=1.5.0
# Testing
pytest>=7.0.0